% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{averageMips}
\alias{averageMips}
\title{averageMips: Average the MIP values}
\usage{
averageMips(filenames, fdir, ...)
}
\arguments{
\item{filenames}{string specifying path and MIP files of interest}

\item{fdir}{specifies where the new file should be placed}

\item{...}{additional arguments}
}
\value{
saves csv, returns path and name of csv with mean MIP values
}
\description{
Find mean (and sd) the posterior MIP values in mip table.
}
\examples{
## not run
}
