% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{batch.plotter.wrapper.new}
\alias{batch.plotter.wrapper.new}
\title{batch.plotter.wrapper.new: Make dot plots by batch}
\usage{
batch.plotter.wrapper.new(data, phenotype, trt.string, ctrl.string, ...)
}
\arguments{
\item{data}{the data frame being used, in this case, from FluDiData}

\item{phenotype}{the phenotype for which you are making plots}

\item{trt.string}{a string indicating the treatment group}

\item{ctrl.string}{a string indicating the control group}

\item{...}{additional arguments}
}
\value{
a wrapper for making pdf plots of treated and control, by batch
}
\description{
Make dot plots, separated by batch (variant).
}
\examples{
## not run
}
