% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{batch.plotter.wrapper}
\alias{batch.plotter.wrapper}
\title{batch.plotter.wrapper: Make dot plots by batch}
\usage{
batch.plotter.wrapper(data, trt.string, ctrl.string, ...)
}
\arguments{
\item{data}{the data frame being used, in this case, from FluDiData}

\item{trt.string}{a string indicating the treatment group}

\item{ctrl.string}{a string indicating the control group}

\item{...}{additional arguments}
}
\value{
a wrapper for making pdf plots of treated and control, by batch
}
\description{
Generate dot plots, separated by batch.
}
\examples{
## not run
}
