% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{batched.plotter}
\alias{batched.plotter}
\title{batched.plotter: Make batched dot plots}
\usage{
batched.plotter(data.frame, title, trt.string, ctrl.string, sexed = c(FALSE,
  TRUE)[1], ...)
}
\arguments{
\item{data.frame}{the data frame containing elements}

\item{title}{the title of the plot}

\item{trt.string}{a string indicating the treatment group}

\item{ctrl.string}{a string indicating the control group}

\item{sexed}{define whether or not to indicate sex-specificity}

\item{...}{additional arguments}
}
\value{
make pdf plots of treated and control, by batch, and by sex
}
\description{
Create batched dot plots indicating treated and control for each jk cross.
}
\examples{
## not run
}
