% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{col.switch}
\alias{col.switch}
\title{col.switch: Convert CC names to colors}
\usage{
col.switch(arg, ...)
}
\arguments{
\item{arg}{the argument, in the form of "AA", "BB", etc.}

\item{...}{additional arguments}
}
\value{
Defines standard CC founder colors in hex-code
}
\description{
Convert CC founder abbreviated strain names to standard CC colors
}
\examples{
col.switch("A", "B", "C", "DD", "EE", "FF")
}
