% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{csv.maker}
\alias{csv.maker}
\title{csv.maker: Save analysis to csv}
\usage{
csv.maker(fname, TreatDi, burnin, ...)
}
\arguments{
\item{fname}{the name of the file path}

\item{TreatDi}{the BayesDiallel object}

\item{burnin}{specify the burnin}

\item{...}{additional arguments}
}
\value{
Save csv files
}
\description{
Save BayesDiallel analysis to csv files.
}
\examples{
## not run
}
