% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{diallelMatrixMakerShortname}
\alias{diallelMatrixMakerShortname}
\title{diallelMatrixMakerShortname}
\usage{
diallelMatrixMakerShortname(data, dam.col.name, sire.col.name,
  batch.col.name = NULL, ...)
}
\arguments{
\item{data}{data frame}

\item{dam.col.name}{dam column name}

\item{sire.col.name}{sire column name}

\item{batch.col.name}{name of batch/random effect column}

\item{...}{additional arguments}
}
\value{
returns diallel incidence matrices
}
\description{
Make design matrices for diallel, based on dam and sire columns with short names.
}
\examples{
## not run
}
