% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{dot.plotter}
\alias{dot.plotter}
\title{dot.plotter: Make dot plots indicating treated and control per cross}
\usage{
dot.plotter(data.frame, title, trt.string, ctrl.string, batch.num = TRUE, ...)
}
\arguments{
\item{data.frame}{the data frame containing elements}

\item{title}{the title of the plot}

\item{trt.string}{a string indicating the treatment group}

\item{ctrl.string}{a string indicating the control group}

\item{batch.num}{the number of the experimental batch}

\item{...}{additional arguments}
}
\value{
make pdf plots of treated and control, by batch
}
\description{
Create plots, with dots, indicating treated and control groups in each jk cross.
}
\examples{
## not run
}
