\name{FluDiData}
\docType{data}
\alias{FluDiData}
\title{Phenotypes from Influenza (Flu) Diallel}
\description{
  This data set gives the weight loss phenotypes, cross, and covariate information
  for the influenza virus mouse diallel experiment.
}
\usage{FluDiData}
\format{A table containing > 1000 individual animals and weight loss across 4 days post-infection.}
\source{The laboratory of Mark Heise at UNC-Chapel Hill}
\references{
To be added.
}
\keyword{datasets}
