% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{hpd.plotter}
\alias{hpd.plotter}
\title{hpd.plotter: Make HPD plots}
\usage{
hpd.plotter(plotdir, chain.object, batched = c(FALSE, TRUE)[1],
  fixed = c(FALSE, TRUE)[1], xlim = c(-10, 10), ...)
}
\arguments{
\item{plotdir}{the directory to save the HPD plots}

\item{chain.object}{the MCMC chain object}

\item{batched}{logical that defines whether to plot batch effect estimates}

\item{fixed}{defines whether to plot fixed effect estimates}

\item{xlim}{gives limits of x-axis}

\item{...}{additional arguments}
}
\value{
make HPD plots and save them to plotdir
}
\description{
Generate highest posterior density plots.
}
\examples{
## not run
}
