% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{imputeMissing}
\alias{imputeMissing}
\title{imputeMissing: Impute missing phenotypes/covariates}
\usage{
imputeMissing(dat, phenotype, covariate, postPred, postPred.cov, effectEst,
  SigmaEst, BlockEst, FixedEst, effectEst.cov, SigmaEst.cov, BlockEst.cov,
  toImpute, toCensor, colsToCopy, numImps, savedir, ...)
}
\arguments{
\item{dat}{original data set}

\item{phenotype}{the phenotype column name}

\item{covariate}{the covariate column name}

\item{postPred}{posterior predictive data for phenotype}

\item{postPred.cov}{posterior predictive data for covariate}

\item{effectEst}{posterior effect estimates for phenotype}

\item{SigmaEst}{posterior sigma estimates for phenotype}

\item{BlockEst}{posterior block estimates for phenotype}

\item{FixedEst}{posterior fixed estimates for phenotype}

\item{effectEst.cov}{posterior effect estimates for covariate}

\item{SigmaEst.cov}{posterior sigma estimates for covariate}

\item{BlockEst.cov}{posterior block estimates for covariate}

\item{toImpute}{categories for imputation}

\item{toCensor}{categories for censoring}

\item{colsToCopy}{data columns from imputation frame to copy}

\item{numImps}{number of timesteps to select from total (# imputed datasets to generate)}

\item{savedir}{the directory the files are saved in}

\item{...}{additional arguments}
}
\value{
generates (n=timestepLen) imputed datasets, returns vector of file names
}
\description{
Impute missing phenotype(s)/covariate, with selective censoring,
             from specific diallel categories based on posteriors.
}
\examples{
## not run
}
