% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{inner.plotter}
\alias{inner.plotter}
\title{inner.plotter: Generate plots for each rep}
\usage{
inner.plotter(AFD, plotdir, xlim = c(-12, 12), ...)
}
\arguments{
\item{AFD}{Diallel object to make plots out of.}

\item{plotdir}{refers to the compound path.}

\item{xlim}{limits for x-axis}

\item{...}{additional arguments}
}
\value{
Makes standard plots in the specified plot directory.
}
\description{
Generate plots from AFD object for each rep.
}
\examples{
## This example may take a couple of minutes
data(FluDiData, envir = environment())
data(PiximusData, envir = environment())
write.csv(FluDiData, file="FluDiData.csv")
filename <- "FluDiData.csv"
args <- c("All", "D0", "NULL", "NULL", "pre", FALSE, "Trt")
treatment 		<- args[1]
phenotype  		<- args[2]
random 			<- args[3]
fixed 			<- args[4]
type				<- args[5]
BS 				<- args[6]
trt.colname      <- args[7]
savedir			<- "."
returned <- run.tr.diallel(filename=filename, savedir=savedir, treatment=treatment, 
trt.colname=trt.colname, phenotype=phenotype, random=random,
fixed=fixed, type=type, BS=BS, thin=1, lengthChains=3000, numChains=5, 
burnin=500, force = TRUE)
plotdir <- returned[[1]]
AFD <- returned[[2]]
inner.plotter(AFD=AFD, plotdir=plotdir)
}
