% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{make.matches}
\alias{make.matches}
\title{make.matches: make matches}
\usage{
make.matches(data, reps, trt.colname = "Trt", trt.string, ctrl.string, fdir,
  force = TRUE, strategy = c("one-to-one", "two-to-one", "three-to-one",
  "mean")[1], select.phenotype = FALSE, ignore.batch = FALSE,
  ignore.day = TRUE, batch = "Block", matchname = "", ...)
}
\arguments{
\item{data}{data set}

\item{reps}{number of repetitions or matches}

\item{trt.colname}{treatment column name}

\item{trt.string}{treated}

\item{ctrl.string}{control}

\item{fdir}{file directory}

\item{force}{logical indicating whether to force matching despite warnings/errors}

\item{strategy}{this gives the treatment-to-control ratio (one-to-one, two-to-one), or mean match}

\item{select.phenotype}{logical indicating whether to select a specific phenotype}

\item{ignore.batch}{logical indicating whether to match independently of batch (TRUE), or within batch (FALSE)}

\item{ignore.day}{logical indicating whether to match independently of day assignment (TRUE), or within day assignment (FALSE)}

\item{batch}{character string giving batch column name}

\item{matchname}{name of match file}

\item{...}{additional arguments}
}
\value{
returns matched treatment response files
}
\description{
Makes one or more matches between treated and control mice in the flu diallel 
based on specified classes and match strategy. In the manuscript, matches do not differ
between imputations (same quartets always selected), but imputed phenotype values differ.
}
\examples{
## This example may take a couple of minutes
data(FluDiData)
write.csv(FluDiData, file="FluDiData.csv")
filename <- "FluDiData.csv"
data <- read.csv("FluDiData.csv")
reps <- 1
trt.string <- "FLU"
ctrl.string <- "MOCK"
fdir <- "." 
strategy <- "mean"
make.matches(data=data, reps=reps, trt.string=trt.string, 
             ctrl.string=ctrl.string, fdir=fdir, strategy=strategy)
}
