% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{mcmc.stack.and.burn}
\alias{mcmc.stack.and.burn}
\title{mcmc.stack.and.burn: Stack MCMC chains, after burning each}
\usage{
mcmc.stack.and.burn(coda.object, burnin, ...)
}
\arguments{
\item{coda.object}{this is the mcmc object}

\item{burnin}{this is the common amount to burn off each chain}

\item{...}{additional arguments}
}
\value{
returns single chain, stacked and burned MCMC object
}
\description{
This is a modification of mcmc.stack that also removes specified number of burn-ins from chains.
}
\examples{
NULL
}
