% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{Mx1HaploClass}
\alias{Mx1HaploClass}
\title{Mx1HaploClass: Convert CC founder strain letter to Mx subspecies haplotype}
\usage{
Mx1HaploClass(arg, ...)
}
\arguments{
\item{arg}{A through H (shorthand for founders)}

\item{...}{additional arguments}
}
\value{
returns dom, mus, cast based on substrain origin
}
\description{
Convert CC founder strain letter to functional Mx1 subspecies haplotype group class.
}
\examples{
## not run
## Note: Ferris et al., 2013, PLoS Pathogens: 
"three functionally distinct classes corresponding to 
## domesticus (dom: A/J, C57BL6/J, 129s1/SvImJ, NOD/HiLtJ and WSB/EiJ), 
## castaneus (cast: CAST/EiJ) and musculus (mus: PWK/PhJ and NZO/ShILtJ)."
}
