% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{plotBatchEffects}
\alias{plotBatchEffects}
\title{plotBatchEffects: Plot batch effects across timecourse}
\usage{
plotBatchEffects(savedirs, batchIDs, batchNames, muID, xlim = c(85, 101), ...)
}
\arguments{
\item{savedirs}{one or more directories containing analysis files}

\item{batchIDs}{the IDs for the batches to be plotted}

\item{batchNames}{the names of the batches to be plotted}

\item{muID}{the ID for the posterior mean}

\item{xlim}{gives limits of x-axis}

\item{...}{additional arguments}
}
\value{
generates timecourse plot and effect estimate plots (ordered, unordered)
}
\description{
Plot batch effects across the timecourse.
}
\examples{
## not run
}
