% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{plotHPDccColors}
\alias{plotHPDccColors}
\title{plotHPDccColors}
\usage{
plotHPDccColors(coda.object, wanted = varnames(coda.object),
  prob.wide = 0.95, prob.narrow = 0.5, xlab = "HPD interval",
  names = NULL, type = "p", name.margin = 6.1, plt.left = NULL,
  plt.right = NULL, plt.bottom = NULL, plt.title = NULL, ylab = "",
  name.line = 3.9, main = "", main.line = 2, col = "black",
  stack = TRUE, ...)
}
\arguments{
\item{coda.object}{The CODA MCMC object}

\item{wanted}{subset of variable names to plot}

\item{prob.wide}{fractional probability for wide interval}

\item{prob.narrow}{fractional probability for narrow interval}

\item{xlab}{x-axis label}

\item{names}{names of variables to use on plot}

\item{type}{type of plot, i.e. "p"}

\item{name.margin}{Margin width}

\item{plt.left}{left plot}

\item{plt.right}{right plot}

\item{plt.bottom}{bottom plot}

\item{plt.title}{title of plot}

\item{ylab}{y-axis label}

\item{name.line}{name line spacing}

\item{main}{main title}

\item{main.line}{main line spacing}

\item{col}{color(s) for HPD}

\item{stack}{logical indicating whether to stack MCMC chains first}

\item{...}{additional arguments}
}
\value{
plots data, using colored HPDs
}
\description{
Plot HPD function, allowing color specification.
}
\examples{
## not run
}
