% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{plotMips}
\alias{plotMips}
\title{plotMips: Plot the MIPs}
\usage{
plotMips(fdir, fname, reorder = FALSE, ...)
}
\arguments{
\item{fdir}{file directory}

\item{fname}{file name}

\item{reorder}{optional reordering of arguments}

\item{...}{additional arguments}
}
\value{
make MIP plots and save them to file directory
}
\description{
Plot model inclusion probabilities, saved as a csv.
}
\examples{
## not run
}
