% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{plotMipsOverMIMPs}
\alias{plotMipsOverMIMPs}
\title{plotMipsOverMimps: Plot MIPs over reps.}
\usage{
plotMipsOverMIMPs(fdir, fname, its, ...)
}
\arguments{
\item{fdir}{file directory}

\item{fname}{file name}

\item{its}{number of MIMPs}

\item{...}{additional arguments}
}
\value{
make MIP plot panel and save them to file directory
}
\description{
Plot model inclusion probabilities over the multiple imputation matched pair/quartets.
}
\examples{
## not run
}
