% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{plotVarps}
\alias{plotVarps}
\title{plotVarps: Plot the variance projections}
\usage{
plotVarps(fdir, fname, reorder = FALSE, ...)
}
\arguments{
\item{fdir}{file directory}

\item{fname}{file name}

\item{reorder}{optional reordering of arguments}

\item{...}{additional arguments}
}
\value{
make variance projection plots and save them to file directory
}
\description{
Plot variance projection (VarP) confidence intervals.
}
\examples{
## not run
}
