% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{poe.analyzer}
\alias{poe.analyzer}
\title{poe.analyzer: Generate plots of POE}
\usage{
poe.analyzer(stacked, plotdir, ...)
}
\arguments{
\item{stacked}{stacked chains}

\item{plotdir}{the directory where the POE plots should be placed}

\item{...}{additional arguments}
}
\value{
saves plots of asymmetric cross-pair-specific effects
}
\description{
Generate plots of estimated parent-of-origin effects.
}
\examples{
#not.run
}
