% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{removeDots}
\alias{removeDots}
\alias{removeDots}
\title{removeDots: remove dots/underscores from a string}
\usage{
removeDots(x)

removeDots(x)
}
\arguments{
\item{x}{A string or vector of strings}

\item{string}{the dot-containing string or a vector of character strings}

\item{...}{additional arguments}

\item{...}{additional arguments}
}
\value{
Returns string without dot/underscore.

Returns string (or vector of strings) without dots or underscores
}
\description{
This removes dots and underscores from a string, replacing them with a space.

Remove dots and underscores from strings
}
\section{treatmentResponseDiallel functions}{

}

\examples{
removeDots("This_is_a.test.string.")
removeDots("This.is.a.test_string")
}
