% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{run.tr.diallel}
\alias{run.tr.diallel}
\title{run.tr.diallel: run tRD analysis.}
\usage{
run.tr.diallel(filename, savedir, treatment, trt.colname, phenotype, random,
  fixed, type, BS = TRUE, thin, lengthChains = 3000, numChains, burnin,
  force = TRUE, ...)
}
\arguments{
\item{filename}{name of data file}

\item{savedir}{directory to save files in}

\item{treatment}{indicate subset}

\item{trt.colname}{indicate treatment columname}

\item{phenotype}{indicate parameter of interest}

\item{random}{name of random effect, or NULL}

\item{fixed}{name of fixed effect, or NULL}

\item{type}{specify matched pairs or other type of analysis}

\item{BS}{whether or not to use BayesSpike}

\item{thin}{thinning factor for chains}

\item{lengthChains}{MCMC chain length}

\item{numChains}{number of independent MCMC chains}

\item{burnin}{amount of burn-in (discarded) samples from warm-up}

\item{force}{logical indicating whether to proceed despite warnings/errors}

\item{...}{additional arguments}
}
\value{
runs the analysis, generating data files.
}
\description{
Run the treatment response diallel analysis.
}
\examples{
## This example may take a couple of minutes
data(FluDiData)
data(PiximusData, envir = environment())
write.csv(FluDiData, file="FluDiData.csv")
filename <- "FluDiData.csv"
args <- c("All", "D0", "NULL", "NULL", "pre", FALSE, "Trt")
treatment 		<- args[1]
phenotype  		<- args[2]
random 			<- args[3]
fixed 			<- args[4]
type				<- args[5]
BS 				<- as.logical(args[6])
trt.colname      <- args[7]
savedir			<- "."
run.tr.diallel(filename=filename, savedir=savedir, treatment=treatment, 
trt.colname=trt.colname, phenotype=phenotype, random=random,
fixed=fixed, type=type, BS=BS, thin=1, lengthChains=3000, numChains=5, 
burnin=500, force = TRUE)
}
