\name{SARSDiData}
\docType{data}
\alias{SARSDiData}
\title{Phenotypes from SARS Diallel}
\description{
  This data set gives the phenotypes, cross, and covariate information
  for the SARS coronavirus mouse diallel experiment.
}
\usage{SARSDiData}
\format{A table containing > 1000 individual animals and dozens of phenotypes}
\source{The laboratories of Ralph Baric and Mark Heise at UNC-Chapel Hill}
\references{
To be added.
}
\keyword{datasets}
