% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{simplify.labels}
\alias{simplify.labels}
\title{simplify.labels: Convert labels}
\usage{
simplify.labels(vector.of.strings, ...)
}
\arguments{
\item{vector.of.strings}{defines the arguments to be substituted, if applicable, with new strings}

\item{...}{additional arguments}
}
\value{
convert effects/parameters to more readable formats
}
\description{
Convert/simplify labels such that they are readable in a table.
}
\examples{
simplify.labels("Gender:motherj")
}
