% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{stackMatchReps}
\alias{stackMatchReps}
\title{stackMatchReps: Combine results across replicates}
\usage{
stackMatchReps(filenames, fdir, ...)
}
\arguments{
\item{filenames}{stack mp imputations based on filenames}

\item{fdir}{file directory}

\item{...}{additional arguments}
}
\value{
saves csv and plot
}
\description{
Stack multiple imputation matched replicates.
}
\examples{
data(FluDiData)
write.csv(FluDiData, file="FluDiData.csv")
filename <- "FluDiData.csv"
data <- read.csv("FluDiData.csv")
reps <- 2
trt.string <- "FLU"
ctrl.string <- "MOCK" 
strategy <- "one-to-one"
fdir <- getwd()
filenames	<- make.matches(data=data, reps=reps, trt.string=trt.string, 
                             ctrl.string=ctrl.string, strategy=strategy, fdir=".")
MIMPstack	<- stackMatchReps(filenames, fdir=fdir)
}
