% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{sub.labels}
\alias{sub.labels}
\title{sub.labels: Replace labels with just the strain information}
\usage{
sub.labels(vector.of.strings, ...)
}
\arguments{
\item{vector.of.strings}{defines the arguments to be substituted, if applicable, with new strings}

\item{...}{additional arguments}
}
\value{
returns vector of strings that contains just the strain information
}
\description{
Replace labels with just the strain information, removing the prefix portion of the string.
}
\examples{
simplify.labels("additive:AJ")
}
