% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\docType{package}
\name{treatmentResponseDiallel}
\alias{treatmentResponseDiallel}
\alias{treatmentResponseDiallel-package}
\title{treatmentResponseDiallel: A package for analysis of infection/treatment response in a diallel cross of inbred lines.
The treatment response diallel (tRD) package provides the important function:
run.tr.diallel}
\description{
treatmentResponseDiallel: A package for analysis of infection/treatment response in a diallel cross of inbred lines.
The treatment response diallel (tRD) package provides the important function:
run.tr.diallel
}
