% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentResponseDiallel.R
\name{var.translate.glmm}
\alias{var.translate.glmm}
\title{var.translate.glmm: translate variable names from BayesDiallel MCMCglmm model}
\usage{
var.translate.glmm(variable.vec, ...)
}
\arguments{
\item{variable.vec}{a vector of the name(s) of variables to be translated}

\item{...}{additional arguments}
}
\value{
returns translation result
}
\description{
Translate BayesDiallel MCMCglmm variable names to standard formats.
}
\examples{
strings <- c("(Intercept)", "is.female", "units")
var.translate.glmm(strings)
}
